+++ oshi-core/src/test/java/oshi/software/os/shared/OperatingSystemTest.java
@@ -169,10 +169,12 @@ class OperatingSystemTest {
         assertThat("Bitness must be 0, 32 or 64", proc.getBitness(), is(oneOf(0, 32, 64)));
         assertThat("Current process open file handles should be -1 or higher", proc.getOpenFiles(),
                 is(greaterThanOrEqualTo(0L)));
-        assertThat("Soft open file limit for process should be -1 or higher", proc.getSoftOpenFileLimit(),
-                is(greaterThanOrEqualTo(-1L)));
-        assertThat("Hard open file limit for process should be -1 or higher", proc.getHardOpenFileLimit(),
-                is(greaterThanOrEqualTo(-1L)));
+        long softOpenFileLimit = proc.getSoftOpenFileLimit();
+        long hardOpenFileLimit = proc.getHardOpenFileLimit();
+        assertThat("Soft open file limit for current process should be positive", softOpenFileLimit,
+                is(greaterThan(0L)));
+        assertThat("Hard open file limit for current process should be at least the soft limit", hardOpenFileLimit,
+                is(greaterThanOrEqualTo(softOpenFileLimit)));
         // updateAttributes should succeed and kernel/user time should be nondecreasing
         long kernelBefore = proc.getKernelTime();
         long userBefore = proc.getUserTime();